refactor: update import paths for SurfaceFactory in surfaces module and test cases
