+++ optiland/surfaces/__init__.py
@@ -4,4 +4,4 @@ from .standard_surface import Surface
 from .object_surface import ObjectSurface
 from .image_surface import ImageSurface
 from .surface_group import SurfaceGroup
-from .factories import SurfaceFactory
+from .factories.surface_factory import SurfaceFactory
